Basic handling of scale dependant layers
This makes it so that layers that are shown/hidden because they are scale
dependant also have another class in the layer tree node.

The code contains somne TODOs we should tackle later. Generally I would
suggest to put some (bigger) part of the LayerTree class into the
LayerTreeNode class.