refactor: rewrite isURL to remove potential security risk
SonarQube marked the regular expresion based implementation of isURL
as a potential security risk:

> Make sure the regex used here, which is vulnerable to super-linear
> runtime due to backtracking, cannot lead to denial of service.
> Using slow regular expressions is security-sensitive

This commit changes the implementation to use URL() object, and it
enforces some rules that are more strict than technically needed. See
the tests for accepted and declined cases. It is quite verbose, but
should remove the flagging as security risk.

The method itself should probably be moved to a generic library, e.g.
https://github.com/terrestris/base-util but I'll leave this for others
to decide.